# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tada",
    "description": "The tada HTTP API. The list of problem codes of each operation is open.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1"
  },
  "paths": {
...
    "/api/v1/events/{event_id}/profile": {
      "get": {
        "tags": [
          "facts"
        ],
        "summary": "Reads the event profile. Each member who can read the event can read it.",
        "operationId": "get_event_profile",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "The ID of the event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The event profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfile"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "organization-required",
          "not-found",
          "unavailable",
          "internal"
        ]
      }
    },
    "/api/v1/events/{event_id}/fields": {
      "get": {
        "tags": [
          "facts"
        ],
        "summary": "Lists the field catalog of an event: the shipped fields and the fields of the event, in the order of their keys.\nEach field has the JSON Schema of its values.",
        "operationId": "list_fields",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "The ID of the event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The field catalog of the event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldPage"
                }
              }
            }
          },
          "default": {
            "description": "A problem (ADR 0037).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-tada-problem-codes": [
          "malformed-request",
          "unauthenticated",
          "organization-required",
          "not-found",
          "unavailable",
          "internal"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Choice": {
        "type": "object",
        "description": "One choice of a choice field.",
        "required": [
          "key",
          "label"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "$ref": "#/components/schemas/Label"
          }
        }
      },
...
      "EventProfile": {
        "type": "object",
        "description": "The event profile: the current facts of the event, its open fact proposals and its open questions.",
        "required": [
          "facts",
          "proposals",
          "open_questions"
        ],
        "properties": {
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "description": "The current version of each fact of the event, in the order of the field keys.\nA field without a fact is absent. An unknown fact is present with the state `unknown`."
          },
          "proposals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FactProposal"
            },
            "description": "The fact proposals of the event without a review result, oldest first. They are not accepted state."
          },
          "open_questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenQuestion"
            },
            "description": "The open questions of the event, in the order of their numbers."
          }
        }
      },
...
      "Evidence": {
        "type": "object",
        "description": "One evidence link: a passage of a source version.",
        "required": [
          "source_version_id",
          "passage"
        ],
        "properties": {
          "source_version_id": {
            "type": "string",
            "format": "uuid"
          },
          "passage": {
            "$ref": "#/components/schemas/Passage"
          }
        }
      },
...
      "Fact": {
        "type": "object",
        "description": "The current version of one fact.",
        "required": [
          "id",
          "field_id",
          "field_key",
          "state",
          "version",
          "evidence"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "field_id": {
            "type": "string",
            "format": "uuid"
          },
          "field_key": {
            "type": "string",
            "description": "The key of the field, for example `date_window`."
          },
          "state": {
            "$ref": "#/components/schemas/FactState"
          },
          "value": {
            "description": "The value. It is absent if the state is `unknown`."
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "description": "The number of the current fact version. A proposal that changes the fact expects it."
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "description": "The passages of source versions that support the current fact version."
          }
        }
      },
      "FactProposal": {
        "type": "object",
        "description": "An open proposal that sets the fact of a field.",
        "required": [
          "id",
          "changeset_id",
          "field_id",
          "state",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "changeset_id": {
            "type": "string",
            "format": "uuid"
          },
          "field_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "$ref": "#/components/schemas/FactState"
          },
          "value": {
            "description": "The proposed value. It is absent if the proposed state is `unknown`."
          },
          "expected_version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The fact version that the proposal expects. It is absent if the field had no fact."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FactState": {
        "type": "string",
        "description": "The status of a fact: `unknown` has no value (ADR 0049).",
        "enum": [
          "accepted",
          "assumption",
          "unknown"
        ]
      },
      "Field": {
        "type": "object",
        "description": "A field definition of the field catalog.",
        "required": [
          "id",
          "key",
          "label",
          "value_type",
          "value_schema",
          "description",
          "module",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string",
            "description": "The stable `snake_case` key, for example `visitor_estimate`."
          },
          "label": {
            "$ref": "#/components/schemas/Label"
          },
          "value_type": {
            "$ref": "#/components/schemas/ValueType"
          },
          "value_schema": {
            "type": "object",
            "description": "The JSON Schema of a value of this field: the `value` of a proposal that sets its fact."
          },
          "description": {
            "type": "string",
            "description": "The meaning of the field, in English."
          },
          "module": {
            "type": "string",
            "description": "The module of the field, for example `core`."
          },
          "status": {
            "$ref": "#/components/schemas/FieldStatus"
          },
          "event_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The event of a field that the event added. It is absent for a shipped field."
          }
        }
      },
      "FieldPage": {
        "type": "object",
        "description": "The field catalog of an event. It has few fields, so the list has one page.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Field"
            }
          }
        }
      },
      "FieldStatus": {
        "type": "string",
        "description": "A deprecated field is readable, but takes no new facts.",
        "enum": [
          "active",
          "deprecated"
        ]
      },
      "Granularity": {
        "type": "string",
        "description": "The meaning of the dates of a date window.",
        "enum": [
          "day",
          "week",
          "month"
        ]
      },
...
      "Label": {
        "oneOf": [
          {
            "type": "object",
            "description": "The ID of a Fluent message of the web client, for example `field-date_window`.",
            "required": [
              "id",
              "kind"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "message"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "text",
              "kind"
            ],
            "properties": {
              "text": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "text"
                ]
              }
            }
          }
        ],
        "description": "The label that people see: a Fluent message of a shipped field, or the German text of a field of an event."
      },
...
      "OpenQuestion": {
        "type": "object",
        "description": "An open question of the event.",
        "required": [
          "id",
          "local_id",
          "text",
          "owner_id",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "local_id": {
            "type": "string",
            "description": "The event-local ID, for example `QST-1` (ADR 0038)."
          },
          "text": {
            "type": "string"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "description": "The user ID of the member who owns the question."
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Passage": {
        "type": "object",
        "description": "A passage of a source version: a range of characters in its normalized text and the exact quote.",
        "required": [
          "start",
          "end",
          "quote"
        ],
        "properties": {
          "start": {
            "type": "integer",
            "format": "int32",
            "description": "The offset of the first character, in Unicode scalar values.",
            "minimum": 0
          },
          "end": {
            "type": "integer",
            "format": "int32",
            "description": "The offset after the last character.",
            "minimum": 0
          },
          "quote": {
            "type": "string"
          },
          "page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The page of a PDF, from 1.",
            "minimum": 0
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "An error response. `detail` never repeats input values.",
        "required": [
          "type",
          "code",
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "The URL of the code in the public catalog."
          },
          "code": {
            "type": "string",
            "description": "The stable problem code. Clients react to this field. The list of codes is open."
          },
          "title": {
            "type": "string",
            "description": "A short, stable English text for developers."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "An English text about this case, for developers. Clients never show it to members."
          },
          "instance": {
            "type": "string",
            "description": "`urn:uuid:` and the request ID."
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemError"
            },
            "description": "The invalid values of a `validation-failed` problem."
          }
        }
      },
      "ProblemError": {
        "type": "object",
        "description": "One invalid value.",
        "required": [
          "pointer",
          "code"
        ],
        "properties": {
          "pointer": {
            "type": "string",
            "description": "A JSON pointer to the value in the request body, for example `/key`."
          },
          "code": {
            "type": "string",
            "description": "The reason, for example `taken`. The list of codes is open."
          }
        }
      },
...
      "ReferenceTarget": {
        "type": "string",
        "description": "The kind of record that a reference points to. The list of kinds is open.",
        "enum": [
          "document",
          "event"
        ]
      },
...
      "ValueType": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "text"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "boolean"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "unit",
              "type"
            ],
            "properties": {
              "unit": {
                "type": "string",
                "description": "The `snake_case` key of the unit, for example `person_per_day`."
              },
              "type": {
                "type": "string",
                "enum": [
                  "quantity"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "currency",
              "type"
            ],
            "properties": {
              "currency": {
                "type": "string",
                "description": "An ISO 4217 code, for example `CHF`."
              },
              "type": {
                "type": "string",
                "enum": [
                  "money"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "date"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "granularity": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Granularity",
                    "description": "Absent if the field accepts each granularity."
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": {
                "type": "string",
                "enum": [
                  "date_window"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "values",
              "multiple",
              "type"
            ],
            "properties": {
              "values": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Choice"
                }
              },
              "multiple": {
                "type": "boolean",
                "description": "True if a fact can hold several choices."
              },
              "type": {
                "type": "string",
                "enum": [
                  "choice"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "target",
              "type"
            ],
            "properties": {
              "target": {
                "$ref": "#/components/schemas/ReferenceTarget"
              },
              "type": {
                "type": "string",
                "enum": [
                  "reference"
                ]
              }
            }
          }
        ],
        "description": "The value type of a field, with its unit, currency or choices. The list of types is open."
      }
    }
  }
}